CLN: refactoring module structure and import statements
