+++ force_gromacs/data_sources/molecule/tests/test_molecule_data_source.py
@@ -3,7 +3,7 @@ from unittest import TestCase, mock
 from force_bdss.api import DataValue
 
 from force_gromacs.gromacs_plugin import GromacsPlugin
-from force_gromacs.tests.probe_classes import (
+from force_gromacs.tests.probe_classes.chemicals import (
     ProbeGromacsFragment, data, mock_method
 )
